+++ backend/app/db.py
@@ -1,0 +1,43 @@
+
+from imports import os
+import mysql.connector
+from config import MYSQL_HOST, MYSQL_USER, MYSQL_PASSWORD, MYSQL_DB
+
+def get_db_connection():
+    try:
+        conn = mysql.connector.connect(
+            host=MYSQL_HOST,
+            user=MYSQL_USER,
+            password=MYSQL_PASSWORD,
+            database=MYSQL_DB,
+            ssl_disabled=True
+        )
+        return conn
+    except mysql.connector.Error as err:
+        print(f"Error: {err}")
+        return None
+
+def initialize_database():
+    """Initializes the database by creating the `users` table if it doesn't exist."""
+    conn = get_db_connection()
+    if conn is None:
+        print("Failed to connect to the database. Cannot initialize.")
+        return
+
+    cursor = conn.cursor()
+    try:
+        cursor.execute("""
+            CREATE TABLE IF NOT EXISTS users (
+                id INT AUTO_INCREMENT PRIMARY KEY,
+                google_id VARCHAR(255) UNIQUE NOT NULL,
+                name VARCHAR(255) NOT NULL,
+                email VARCHAR(255) UNIQUE NOT NULL
+            )
+        """)
+        conn.commit()
+        print("Database initialized successfully.")
+    except mysql.connector.Error as err:
+        print(f"Error initializing database: {err}")
+    finally:
+        cursor.close()
+        conn.close()
